+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Damage/Tabs/DamageDoneByPartyMembers/Tabs/DamageDoneByPartyMember.lua
@@ -9,10 +9,12 @@ UI
 
 local L = LibStub("AceLocale-3.0"):GetLocale("MyDungeonsBook");
 
-
-local function getDamageByPartyMemberSpellMenu(spellId)
+local function getDamageByPartyMemberSpellMenu(rows, index, cols, challengeId, unitId)
+    local spellId = rows[index].cols[1].value;
+    local report = MyDungeonsBook:DamageDoneByPartyMemberFrame_Report_Create(rows[index], cols, challengeId, unitId);
     return {
-        MyDungeonsBook:WowHead_Menu_Spell(spellId)
+        MyDungeonsBook:WowHead_Menu_Spell(spellId),
+        MyDungeonsBook:Report_Menu(report)
     };
 end
 
@@ -33,7 +35,7 @@ function MyDungeonsBook:DamageDoneByPartyMemberFrame_Create(parentFrame, challen
             if (button == "RightButton" and realrow) then
                 local spellId = data[realrow].cols[1].value;
                 if (spellId > 0) then
-                    EasyMenu(getDamageByPartyMemberSpellMenu(spellId), self.menuFrame, "cursor", 0 , 0, "MENU");
+                    EasyMenu(getDamageByPartyMemberSpellMenu(data, realrow, table.cols, challengeId, unitId), self.menuFrame, "cursor", 0 , 0, "MENU");
                 end
             end
         end,
@@ -289,4 +291,18 @@ function MyDungeonsBook:DamageDoneByPartyMemberFrame_GetDataForTable(challengeId
     end
     tinsert(tableData, summaryRow);
     return tableData, summaryRow;
+end
+
+--[[--
+@param[type=table] row
+@param[type=table] cols
+@param[type=number] challengeId
+@param[type=unitId] unitId
+@return[type=table]
+]]
+function MyDungeonsBook:DamageDoneByPartyMemberFrame_Report_Create(row, cols, challengeId, unitId)
+    local spellLink = GetSpellLink(row.cols[1].value);
+    local challenge = self:Challenge_GetById(challengeId);
+    local title = string.format(L["MyDungeonsBook Damage stats by %s with %s:"], challenge.players[unitId].name or "", spellLink);
+    return self:Table_PlayersRowBySpell_Report_Create(row, cols, title);
 end